+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## unreleased
 * Rename `Card` to `CardPayments`
+* PayPalUI
+    * Fix issue where label was not being shown
+>>>>>>> main
 
 ## 0.0.4 (2023-01-17)
 * Card